+++ src/App.tsx
@@ -1,10 +1,13 @@
 import React, { useState } from 'react'
 
-import { DeleteConfirmation } from './DeleteConfirmation'
+import { DeleteConfirmation } from './components/DeleteConfirmation'
 import './App.css'
+import { InputTodo } from './components/InputTodo'
+import { TodoStatusCount } from './components/TodoStatusCount'
+import { TodoList } from './components/TodoList'
 
 type Todo = {
-  id: number
+  id: string
   title: string
   completed: boolean
   editing: boolean
@@ -14,120 +17,32 @@ export default function App() {
   const [todoText, setTodoText] = useState('')
   const [todos, setTodos] = useState<Todo[]>([])
   const [showModal, setShowModal] = useState(false)
-  const [deleteTodoId, setDeleteTodoId] = useState<number | null>(null)
-
-  const onChangeTodoText = (e: React.ChangeEvent<HTMLInputElement>) =>
-    setTodoText(e.target.value)
-  const onClickAddTodo = () => {
-    if (todoText === '') return
-    const newTodo: Todo = {
-      id: todos.length + 1,
-      title: todoText,
-      completed: false,
-      editing: false,
-    }
-    setTodos([...todos, newTodo])
-    setTodoText('')
-  }
-  const onChangeTodoCompleted = (todo: Todo) => {
-    const newTodos: Todo[] = todos.slice().map((t) => {
-      if (t.id === todo.id) {
-        t.completed = !t.completed
-      }
-      return t
-    })
-    setTodos(newTodos)
-  }
-  const onClickEditTodo = (todo: Todo) => {
-    const newTodos: Todo[] = todos.slice().map((t) => {
-      if (t.id === todo.id && t.title !== '') {
-        t.editing = !t.editing
-      }
-      return t
-    })
-    setTodos(newTodos)
-  }
-  const onChangeEditTodo = (
-    e: React.ChangeEvent<HTMLInputElement>,
-    todo: Todo
-  ) => {
-    const newTodos: Todo[] = todos.slice().map((t) => {
-      if (t.id === todo.id) {
-        t.title = e.target.value
-      }
-      return t
-    })
-    setTodos(newTodos)
-  }
-  const onClickDeleteConfirm = (todo: Todo) => {
-    setDeleteTodoId(todo.id)
-    setShowModal(true)
-  }
-  const onClickDeleteCancel = () => {
-    setShowModal(false)
-    setDeleteTodoId(null)
-  }
-  const onClickDelete = (deleteTodoId: number) => {
-    const newTodos: Todo[] = todos.filter((t) => t.id !== deleteTodoId).slice()
-    setTodos(newTodos)
-    setShowModal(false)
-    setDeleteTodoId(null)
-  }
+  const [deleteTodoId, setDeleteTodoId] = useState<string>('')
 
   return (
     <>
-      <div className="input-area">
-        <input value={todoText} onChange={onChangeTodoText} />
-        <button onClick={onClickAddTodo}>保存</button>
-      </div>
+      <InputTodo
+        todos={todos}
+        todoText={todoText}
+        setTodoText={setTodoText}
+        setTodos={setTodos}
+      />
       <div className="todo-area">
         <p className="title">TODO</p>
-        <div>
-          <p>全てのタスク：{todos.length}</p>
-          <p>完了済み：{todos.filter((todo) => todo.completed).length}</p>
-          <p>全てのタスク：{todos.length}</p>
-        </div>
-        <ul>
-          {todos.map((todo) => {
-            return (
-              <li key={todo.id} className="list-row">
-                <input
-                  type="checkbox"
-                  checked={todo.completed}
-                  onChange={() => onChangeTodoCompleted(todo)}
-                />
-                {todo.editing ? (
-                  <>
-                    <input
-                      value={todo.title}
-                      onChange={(e) => onChangeEditTodo(e, todo)}
-                    />
-                    <button onClick={() => onClickEditTodo(todo)}>保存</button>
-                  </>
-                ) : (
-                  <>
-                    <p>{todo.title}</p>
-                    {todos.filter((t) => t.editing).length === 0 && (
-                      <>
-                        <button onClick={() => onClickEditTodo(todo)}>
-                          編集
-                        </button>
-                        <button onClick={() => onClickDeleteConfirm(todo)}>
-                          削除
-                        </button>
-                      </>
-                    )}
-                  </>
-                )}
-              </li>
-            )
-          })}
-        </ul>
+        <TodoStatusCount todos={todos} />
+        <TodoList
+          todos={todos}
+          setTodos={setTodos}
+          setShowModal={setShowModal}
+          setDeleteTodoId={setDeleteTodoId}
+        />
         <DeleteConfirmation
+          todos={todos}
           showModal={showModal}
           deleteTodoId={deleteTodoId}
-          onClickDeleteCancel={onClickDeleteCancel}
-          onClickDelete={onClickDelete}
+          setTodos={setTodos}
+          setShowModal={setShowModal}
+          setDeleteTodoId={setDeleteTodoId}
         />
       </div>
     </>
